Replace `RefreshTokenLayer` with `AsyncFilterLayer`
No need to add `#[doc(hidden)]` because they're not accessible and
doesn't show up in docs.
