+++ kube-client/Cargo.toml
@@ -63,7 +63,7 @@ hyper = { version = "0.14.13", optional = true, features = ["client", "http1", "
 hyper-tls = { version = "0.5.0", optional = true }
 hyper-rustls = { version = "0.23.0", optional = true }
 tokio-tungstenite = { version = "0.16.1", optional = true }
-tower = { version = "0.4.6", optional = true, features = ["buffer", "util"] }
+tower = { version = "0.4.6", optional = true, features = ["buffer", "filter", "util"] }
 tower-http = { version = "0.2.0", optional = true, features = ["auth", "map-response-body", "trace"] }
 hyper-timeout = {version = "0.4.1", optional = true }
 tame-oauth = { version = "0.6.0", features = ["gcp"], optional = true }
